+++ src/main/java/pl/thedeem/intellij/dql/services/ui/TenantServiceGroup.java
@@ -1,10 +1,23 @@
 package pl.thedeem.intellij.dql.services.ui;
 
+import com.intellij.icons.AllIcons;
 import com.intellij.navigation.ItemPresentation;
+import com.intellij.openapi.actionSystem.ActionGroup;
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.actionSystem.DefaultActionGroup;
+import com.intellij.openapi.fileEditor.FileEditorManager;
+import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import pl.thedeem.intellij.common.StandardItemPresentation;
+import pl.thedeem.intellij.common.services.ManagedService;
 import pl.thedeem.intellij.common.services.ManagedServiceGroup;
+import pl.thedeem.intellij.common.services.ProjectServicesManager;
+import pl.thedeem.intellij.dql.DQLBundle;
 import pl.thedeem.intellij.dql.DQLIcon;
+import pl.thedeem.intellij.dql.actions.OpenDQLQueryConsoleAction;
+import pl.thedeem.intellij.dql.fileProviders.DQLQueryConsoleVirtualFile;
 import pl.thedeem.intellij.dql.settings.tenants.DynatraceTenant;
 import pl.thedeem.intellij.dql.settings.tenants.DynatraceTenantsService;
 
@@ -12,6 +25,7 @@ import java.util.Objects;
 
 public class TenantServiceGroup implements ManagedServiceGroup {
     private final String tenantId;
+    private DefaultActionGroup actions;
 
     public TenantServiceGroup(@NotNull String tenantId) {
         this.tenantId = tenantId;
@@ -37,5 +51,51 @@ public class TenantServiceGroup implements ManagedServiceGroup {
     @Override
     public int hashCode() {
         return Objects.hashCode(tenantId);
+    }
+
+    @Override
+    public @Nullable ActionGroup getToolbarActions() {
+        if (actions == null) {
+            actions = new DefaultActionGroup();
+            actions.add(new AnAction(
+                    DQLBundle.message("services.tenantGroup.actions.openConsole.title"),
+                    null,
+                    AllIcons.Debugger.ExecuteCurrentStatement
+            ) {
+                @Override
+                public void actionPerformed(@NotNull AnActionEvent e) {
+                    Project project = e.getProject();
+                    if (project == null) {
+                        return;
+                    }
+                    DQLQueryConsoleVirtualFile vf = new DQLQueryConsoleVirtualFile(
+                            DQLBundle.message(
+                                    "action.DQL.OpenDQLQueryConsole.ServiceViewAction.consoleName",
+                                    OpenDQLQueryConsoleAction.COUNTER.incrementAndGet()
+                            ),
+                            tenantId
+                    );
+                    FileEditorManager.getInstance(project).openFile(vf, true);
+                }
+            });
+            actions.addAction(new AnAction(
+                    DQLBundle.message("services.tenantGroup.actions.closeAll.title"),
+                    null,
+                    AllIcons.Actions.Close
+            ) {
+                @Override
+                public void actionPerformed(@NotNull AnActionEvent e) {
+                    Project project = e.getProject();
+                    if (project == null) {
+                        return;
+                    }
+                    ProjectServicesManager manager = ProjectServicesManager.getInstance(project);
+                    for (ManagedService service : manager.find(s -> s.getParentGroups().contains(TenantServiceGroup.this))) {
+                        manager.unregisterService(service);
+                    }
+                }
+            });
+        }
+        return actions;
     }
 }